+++ libbpf-rs/CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased
 - Added `Map::is_pinned` and `Map::get_pin_path` methods for getting map pin status
 - Added `Program::attach_iter` for attaching of programs to an iterator
 - Added `Map::delete_batch` method for bulk deletion of elements
+- Added read/update/delete support for queue and stack `Map` types
 - Improved `btf_type_match!` macro, adding support for most of Rust's `match`
   capabilities
 - Added `skel` module exposing skeleton related traits